+++ projectroles/tests/test_views.py
@@ -3322,6 +3322,25 @@ class TestRoleAssignmentCreateView(
             )
         self.assertEqual(response.status_code, 302)
 
+    def test_get_promote_delegate_limit_reached_inherited_guest(self):
+        """Test GET with inherited guest and delegate limit reached"""
+        user_delegate = self.make_user('user_delegate')
+        self.make_assignment(self.project, user_delegate, self.role_delegate)
+        guest_as_cat = self.make_assignment(
+            self.category, self.user_new, self.role_guest
+        )
+        with self.login(user_delegate):
+            response = self.client.get(
+                reverse(
+                    'projectroles:role_create_promote',
+                    kwargs={
+                        'project': self.project.sodar_uuid,
+                        'promote_as': guest_as_cat.sodar_uuid,
+                    },
+                )
+            )
+        self.assertEqual(response.status_code, 200)
+
     def test_post(self):
         """Test RoleAssignmentCreateView POST"""
         self.assertEqual(RoleAssignment.objects.all().count(), 2)